add bootscene
