+++ src/Scene2.js
@@ -1,0 +1,546 @@
+export default class Scene2 extends Phaser.Scene {
+  constructor() {
+    super('playGame');
+  }
+  preload () {
+
+  }
+  create () {
+    music = this.sound.add('rudebuster');
+    bossMusic = this.sound.add('orchestrabuster');
+    music.play({loop: true, volume: 0.1});
+
+    bgBack = this.add.tileSprite(0, 0, config.width, config.height, 'background4');
+    bgBack.setOrigin(0,0);
+    bgBack.setScrollFactor(0);
+    bgBack.setScale(6);
+
+    bgMid1 = this.add.tileSprite(0, 0, config.width, config.height, 'background3');
+    bgMid1.setOrigin(0,0);
+    bgMid1.setScrollFactor(0);
+    bgMid1.setScale(6);
+
+    bgMid2 = this.add.tileSprite(0, 0, config.width, config.height, 'background2');
+    bgMid2.setOrigin(0,0);
+    bgMid2.setScrollFactor(0);
+    bgMid2.setScale(6);
+
+    bgFront = this.add.tileSprite(0, 0, config.width, config.height, 'background1');
+    bgFront.setOrigin(0,0);
+    bgFront.setScrollFactor(0);
+    bgFront.setScale(6);
+
+    scoreText = this.add.text(16, 360, 'Score: 0', { fontSize: '32px', fill: '#fff' });
+    healthText = this.add.text(16, 420, 'Health: 0', { fontSize: '32px', fill: '#fff' });
+
+    this.cameras.main.setBounds(0, 0, 7700, 950);
+
+    map = this.make.tilemap({ key: 'map' });
+    tileset = map.addTilesetImage('tileset1', 'tiles');
+    platforms = map.createStaticLayer('GroundLayer', tileset, 0, 200);
+    platforms.setScale(.3);
+
+    platforms.setCollisionByExclusion(-1, true);
+
+    player = this.physics.add.sprite(100, 450, 'robot');
+    player.health = 100;
+    boss = this.physics.add.sprite(7380, 300, 'bossidle');
+    boss.setGravityY(-300);
+    boss.setScale(1.3);
+    boss.health = 50;
+
+    coins = this.physics.add.group({
+      key: 'coin',
+      repeat: 200,
+      setXY: { x:100, y:0, stepX: 100 }
+    });
+    coins.children.iterate(function (child) {
+      child.setScale(.2);
+      child.setBounce(0.5);
+    });
+
+    fireBullets = this.physics.add.group();
+
+    whiteSlimes = this.physics.add.group({
+      key: 'whiteslime',
+      repeat: 10,
+      setXY: { x:225, y:0, stepX: 555 }
+    });
+    whiteSlimes.children.iterate(function (child) {
+      child.health = 3;
+      child.setScale(1.5);
+      child.setSize(2, 8, true);
+    });
+
+    redSlimes = this.physics.add.group({
+      key: 'redslime',
+      repeat: Phaser.Math.Between(20, 25),
+      setXY: { x:300, y:0, stepX: Phaser.Math.Between(200, 600), stepY: Phaser.Math.Between(0, 10) }
+    });
+    redSlimes.children.iterate(function (child) {
+      child.health = 3;
+      child.setScale(2);
+      child.setSize(4, 10, true);
+    });
+
+    this.time.addEvent({
+      delay: 5000,
+      callback: () => {
+        redSlimes.children.iterate(function (child) {
+          child.setVelocityY(-300);
+        });
+      },
+      loop: true
+    });
+
+    fireSkulls = this.physics.add.group();
+
+    this.time.addEvent({
+      delay: 2500,
+      callback: () => {
+        fireSkulls.create(Phaser.Math.Between(2000, 6300), Phaser.Math.Between(0, 900), 'skull');
+        fireSkulls.children.iterate(function (child) {
+          child.setGravityY(-300);
+          child.setScale(.7);
+          child.setVelocityX(-160);
+          child.setSize(4, 8, true);
+          child.anims.play('skullfly', true);
+          child.health = 2;
+        });
+      },
+      loop: true
+    });
+
+    greenSlimes = this.physics.add.group();
+    littleSlimes = this.physics.add.group();
+    bossBullets = this.physics.add.group();
+
+    this.time.addEvent({
+      delay: 10000,
+      callback: () => {
+        greenSlimes.create(Phaser.Math.Between(1500, 6000), Phaser.Math.Between(500, 900), 'greenslime');
+        greenSlimes.children.iterate(function (child) {
+            child.setScale(1.7);
+            child.setGravityY(-305);
+            child.anims.play('floatslime', true);
+            child.setSize(4, 4, true);
+            child.health = 5;
+        });
+      },
+      loop: true
+    });
+
+    this.time.addEvent({
+      delay: 6000,
+      callback: () => {
+        greenSlimes.children.iterate(function (child) {
+          littleSlimes.create(child.x, child.y, 'littleslime');
+          littleSlimes.children.iterate(function (child) {
+            child.setScale(1.3);
+            child.setBounce(0.5);
+            child.setSize(1, 1, true)
+            child.health = 1;
+          })
+        });
+      },
+      loop: true
+    });
+
+    this.time.addEvent({
+      delay: 3000,
+      callback: () => {
+          littleSlimes.children.iterate(function (child) {
+            child.setVelocityY(-300);
+            child.setVelocityX(-30);
+          });
+      },
+      loop: true
+    });
+
+
+    player.setBounce(0.2);
+    player.body.setGravityY(300);
+
+    this.cameras.main.startFollow(player, true);
+
+    this.anims.create({
+      key: 'bossbullet',
+      frames: this.anims.generateFrameNumbers('bossbullet',
+        {start: 0, end: 6 }),
+      frameRate: 5,
+      yoyo: true,
+      repeat: -1
+    });
+
+    this.anims.create({
+      key: 'bossattack',
+      frames: this.anims.generateFrameNumbers('bossattack',
+       {start: 0, end: 7}),
+       frameRate: 20,
+    });
+
+    this.anims.create({
+      key: 'bossidle',
+      frames: this.anims.generateFrameNumbers('bossidle',
+        {start: 0, end: 5}),
+      frameRate: 5,
+      repeat: -1
+    });
+
+    this.anims.create({
+      key: 'skullfly',
+      frames: this.anims.generateFrameNumbers('skull',
+        {start: 0, end: 7}),
+      frameRate: 10,
+      repeat: -1
+    })
+
+    this.anims.create({
+      key: 'walkslime',
+      frames: this.anims.generateFrameNumbers('whiteslime',
+        { start: 0, end: 15}
+      ),
+      frameRate: 6,
+      repeat: -1
+    })
+
+    this.anims.create({
+      key: 'jumpslime',
+      frames: this.anims.generateFrameNumbers('redslime',
+        { start: 0, end: 15}
+      ),
+      frameRate: 6,
+      repeat: -1
+    })
+
+    this.anims.create({
+      key: 'floatslime',
+      frames: this.anims.generateFrameNumbers('greenslime',
+        { start: 0, end: 15}
+      ),
+      frameRate: 12,
+      repeat: -1
+    })
+
+    this.anims.create({
+      key: 'fireshoot',
+      frames: this.anims.generateFrameNumbers('firebullet',
+        {start: 0, end: 8}
+      ),
+      frameRate: 5
+    })
+
+    this.anims.create({
+      key: 'coinidle',
+      frames:  this.anims.generateFrameNumbers('coin',
+        {start: 0, end: 7}
+      ),
+      frameRate: 5,
+      repeat: -1
+    })
+
+    this.anims.create({
+      key: 'idle',
+      frames: this.anims.generateFrameNumbers('robot',
+        {start: 0, end: 6}
+      ),
+      frameRate: 3,
+      repeat: -1
+    });
+
+    this.anims.create({
+      key: 'walk',
+      frames: this.anims.generateFrameNumbers('robot',
+        {start: 8, end: 14}
+      ),
+      frameRate: 5,
+      repeat: -1
+    });
+
+    this.anims.create({
+      key: 'shoot',
+      frames: this.anims.generateFrameNumbers('robot',
+      {start: 16, end: 20}
+    ),
+    frameRate: 30,
+    })
+
+    this.anims.create({
+      key: 'jump',
+      frames: this.anims.generateFrameNumbers('robot',
+      {start: 32, end: 36}),
+      frameRate: 10,
+      repeat: -1
+    });
+
+    this.anims.create({
+      key: 'death',
+      frames: this.anims.generateFrameNumbers('robot',
+      {start: 35, end: 44 }),
+      frameRate: 5,
+    });
+
+    cursors = this.input.keyboard.createCursorKeys();
+    this.physics.add.collider(player, platforms);
+    this.physics.add.collider(coins, platforms);
+    this.physics.add.collider(boss, platforms);
+    this.physics.add.collider(bossBullets, platforms, destroyBullet, null, this);
+    this.physics.add.collider(whiteSlimes, platforms);
+    this.physics.add.collider(redSlimes, platforms);
+    this.physics.add.collider(littleSlimes, platforms);
+
+    this.physics.add.overlap(whiteSlimes, fireBullets, damageEnemy, null, this);
+    this.physics.add.overlap(redSlimes, fireBullets, damageEnemy, null, this);
+    this.physics.add.overlap(greenSlimes, fireBullets, damageEnemy, null, this);
+    this.physics.add.overlap(littleSlimes, fireBullets, damageEnemy, null, this);
+    this.physics.add.overlap(fireSkulls, fireBullets, damageEnemy, null, this);
+    this.physics.add.collider(boss, fireBullets, damageBoss, null, this);
+
+    this.physics.add.overlap(player, whiteSlimes, playerDamage5, null, this);
+    this.physics.add.overlap(player, redSlimes, playerDamage10, null, this);
+    this.physics.add.overlap(player, littleSlimes, playerDamage5Disable, null, this);
+    this.physics.add.overlap(player, fireSkulls, playerDamage10Disable, null, this);
+    this.physics.add.overlap(player, greenSlimes, playerDamage10, null, this);
+    this.physics.add.overlap(player, bossBullets, playerDamage10Disable, null, this);
+
+    this.physics.add.collider(fireBullets, platforms, destroyBullet, null, this);
+    this.physics.add.overlap(player, coins, collectCoin, null, this);
+
+    function damageEnemy(enemy, bullet) {
+        enemy.health-=1;
+        enemy.setTint(0xff0000);
+        this.time.addEvent({
+          delay: 200,
+          callback: () => {
+            enemy.setTint(0xffffff);
+          },
+        });
+        bullet.destroy();
+        if (enemy.health <= 0){
+          score += 5;
+          enemy.destroy();
+          coins.create(enemy.x, enemy.y, 'coin');
+          coins.children.iterate( function (child) {
+            child.setScale(.2);
+            child.anims.play('coinidle', true);
+          });
+        }
+    }
+
+    function damageBoss(boss, bullet) {
+      boss.health-=1;
+      boss.setTint(0xff0000);
+      console.log(boss.health);
+      this.time.addEvent({
+        delay: 200,
+        callback: () => {
+          boss.setTint(0xffffff);
+        },
+      });
+      bullet.destroy();
+      if (boss.health === 47){
+        music.pause();
+        bossMusic.play({loop:true, volume: 0.1});
+      }
+      if (boss.health <= 0){
+        boss.destroy();
+        this.physics.pause();
+        score += 200;
+      }
+    }
+
+    function playerDamage10(player, enemy) {
+      player.x-=20;
+      score -= 50;
+      player.setTint(0xff0000);
+      player.health -= 10;
+      console.log(player.health);
+      this.time.addEvent({
+        delay: 100,
+        callback: () => {
+          player.x-=25;
+          player.setTint(0xffffff);
+        },
+      });
+      if (player.health <= 0){
+        score -= 1000;
+        gameOver = true;
+        console.log('game over');
+      }
+    }
+
+    function playerDamage5(player, enemy) {
+      player.x-=15;
+      player.setTint(0xff0000);
+      player.health -= 5;
+      score -= 50;
+      console.log(player.health);
+      this.time.addEvent({
+        delay: 200,
+        callback: () => {
+          player.x-=15;
+          player.setTint(0xffffff);
+        },
+      });
+      if (player.health <= 0){
+        gameOver = true;
+        console.log('game over');
+      }
+    }
+
+    function playerDamage10Disable(player, enemy) {
+      player.x-=15;
+      player.setTint(0xff0000);
+      player.health -= 10;
+      score -= 50;
+      enemy.destroy();
+      console.log(player.health);
+      this.time.addEvent({
+        delay: 200,
+        callback: () => {
+          player.x-=15;
+          player.setTint(0xffffff);
+        },
+      });
+      if (player.health <= 0){
+        gameOver = true;
+        console.log('game over');
+        console.log(score);
+      }
+    }
+
+    function playerDamage5Disable(player, enemy) {
+      player.x-=15;
+      player.setTint(0xff0000);
+      player.health -= 5;
+      score -= 50;
+      enemy.destroy();
+      console.log(player.health);
+      this.time.addEvent({
+        delay: 200,
+        callback: () => {
+          player.x-=15;
+          player.setTint(0xffffff);
+        },
+      });
+      if (player.health <= 0){
+        gameOver = true;
+        console.log('game over');
+      }
+    }
+
+    function collectCoin (player, coin) {
+      coin.disableBody(true, true);
+      score += 10;
+    }
+
+    function destroyBullet(bullet, collider) {
+      bullet.disableBody(true, true);
+    }
+
+    this.anims.staggerPlay('coinidle', coins.getChildren(), 0.03);
+    this.anims.staggerPlay('walkslime', whiteSlimes.getChildren(), 0.03);
+    this.anims.staggerPlay('jumpslime', redSlimes.getChildren(), 0.03);
+    this.anims.staggerPlay('skullfly', fireSkulls.getChildren(), 0.03);
+
+    this.time.addEvent({
+        delay: 528,
+        callback: () => {
+          boss.anims.play('bossattack');
+          var bullet = bossBullets.create(boss.x, boss.y, 'bossbullet');
+          bossBullets.children.iterate( function (child) {
+            child.setScale(.3);
+            child.setGravityY(-290);
+            child.setVelocityX(-200);
+            child.setBounce(.8);
+            child.setSize(4, 4, true)
+            child.anims.play('bossbullet', true);
+          });
+        },
+        loop: true
+      });
+    bossTween = this.tweens.add({
+            targets: boss,
+            props: {
+              y: {value: '+=500', duration: 2125}
+            },
+            ease: 'sine.easeInOut',
+            yoyo: true,
+            repeat: -1
+          });
+    whiteSlimeTween = this.tweens.add({
+      targets: whiteSlimes.getChildren(),
+      props: {
+        x: {value: '+=85', duration: 1500}
+      },
+      ease: 'Sine.easeInOut',
+      yoyo: true,
+      repeat: -1
+    });
+  }
+
+  update () {
+    if(healthText.x >= player.x + 550) {
+      scoreText.x = player.x - 700;
+      healthText.x = player.x - 700;
+    }
+
+    scoreText.setText('Score: ' + score);
+    healthText.setText('Health: ' + player.health);
+
+    //------------------------
+    //------------------------
+
+    var cam = this.cameras.main;
+      this.anims.staggerPlay('fireshoot', fireBullets.getChildren(), 0.03);
+    if (gameOver) {
+      return;
+    }
+
+    if (cursors.left.isDown) {
+      player.setVelocityX(-160);
+      player.anims.play('walk', true);
+      cam.scrollX -= 4;
+      scoreText.x-=2.5;
+      healthText.x-=2.5;
+      bgMid1.tilePositionX -= 0.1;
+      bgMid2.tilePositionX -= 0.2;
+      bgFront.tilePositionX -= 0.3;
+    } else if (cursors.right.isDown) {
+      player.setVelocityX(160);
+      player.anims.play('walk', true);
+      cam.scrollX += 4;
+      scoreText.x+= 4;
+      healthText.x+= 4;
+      bgMid1.tilePositionX += 0.1;
+      bgMid2.tilePositionX += 0.2;
+      bgFront.tilePositionX += 0.3;
+    } else {
+      player.setVelocityX(0);
+      player.anims.play('idle', true);
+    }
+
+    if (cursors.up.isDown && player.body.onFloor()) {
+      player.setVelocityY(-550);
+      player.anims.play('jump', true);
+    }
+
+    if (this.input.keyboard.checkDown(cursors.space, 300)) {
+      console.log(player.x);
+      var shoot = fireBullets.create(player.x + 40, player.y + 15, 'firebullet');
+      if (cursors.left.isDown){
+        shoot.setVelocityX(-500);
+      } else {
+        shoot.setVelocityX(500);
+      }
+      shoot.body.setGravityY(-300);
+    }
+
+    if (player.body.velocity.x > 0) {
+      player.setFlipX(false);
+    } else if (player.body.velocity.x < 0) {
+      player.setFlipX(true);
+    }
+
+  }
+}
