+++ src/game.js
@@ -1,0 +1,19 @@
+var config = {
+  type: Phaser.AUTO,
+  width: 1024,
+  height: 600,
+  physics: {
+    default: 'arcade',
+    arcade: {
+      gravity: { y: 300 },
+      debug: false
+    }
+  },
+  scene: {
+    preload: preload,
+    create: create,
+    update: update
+  }
+};
+
+var game = new Phaser.Game(config);
